Only show systemd unit file with filled-in parameters.
