+++ src/bin/pg_autoctl/cli_do_root.c
@@ -209,7 +209,6 @@ CommandLine *do_subcommands[] = {
 	&do_primary_,
 	&do_standby_,
 	&do_show_commands,
-	&do_systemd_commands,
 	&do_discover,
 	&do_destroy,
 	NULL
