+++ src/bin/pg_autoctl/cli_root.c
@@ -37,6 +37,7 @@ CommandLine *show_subcommands[] = {
 	&show_uri_command,
 	&show_events_command,
 	&show_state_command,
+	&systemd_cat_service_file_command,
 	NULL
 };
